Native version test
