+++ integration-tests/src/test/java/org/jboss/fuse/mvnd/it/VersionTest.java
@@ -15,39 +15,10 @@
  */
 package org.jboss.fuse.mvnd.it;
 
-import java.io.IOException;
-import javax.inject.Inject;
-import org.assertj.core.api.Assertions;
-import org.jboss.fuse.mvnd.assertj.MatchInOrderAmongOthers;
-import org.jboss.fuse.mvnd.client.Client;
-import org.jboss.fuse.mvnd.client.ClientLayout;
-import org.jboss.fuse.mvnd.client.ClientOutput;
 import org.jboss.fuse.mvnd.junit.MvndTest;
-import org.junit.jupiter.api.Test;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Mockito;
+import org.jboss.fuse.mvnd.junit.MvndTestExtension;
 
 @MvndTest(projectDir = "src/test/projects/single-module")
-public class VersionTest {
+public class VersionTest extends VersionNativeIT {
 
-    @Inject
-    Client client;
-
-    @Inject
-    ClientLayout layout;
-
-    @Test
-    void version() throws IOException, InterruptedException {
-        final ClientOutput output = Mockito.mock(ClientOutput.class);
-
-        client.execute(output, "-v").assertSuccess();
-
-        final ArgumentCaptor<String> logMessage = ArgumentCaptor.forClass(String.class);
-        Mockito.verify(output, Mockito.atLeast(1)).accept(logMessage.capture());
-
-        Assertions.assertThat(logMessage.getAllValues())
-                .is(new MatchInOrderAmongOthers<>(
-                        "\\QMaven Daemon " + System.getProperty("project.version") + "\\E",
-                        "\\QMaven home: " + layout.mavenHome() + "\\E"));
-    }
 }
